🚚 api: move repository to src/certification/configuration context
