+++ api/src/certification/configuration/domain/usecases/index.js
@@ -4,14 +4,17 @@ import { fileURLToPath } from 'node:url';
 import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
 import * as complementaryCertificationRepository from '../../../complementary-certification/infrastructure/repositories/complementary-certification-repository.js';
+import * as attachableTargetProfileRepository from '../../infrastructure/repositories/attachable-target-profiles-repository.js';
 
 /**
  *
  * Using {@link https://jsdoc.app/tags-type "Closure Compiler's syntax"} to document injected dependencies
  *
  * @typedef {complementaryCertificationRepository} ComplementaryCertificationRepository
+ * @typedef {attachableTargetProfileRepository} AttachableTargetProfileRepository
  **/
 const dependencies = {
+  attachableTargetProfileRepository,
   complementaryCertificationRepository,
 };
