+++ api/src/certification/configuration/infrastructure/repositories/attachable-target-profiles-repository.js
@@ -1,6 +1,6 @@
-import { knex } from '../../../db/knex-database-connection.js';
+import { knex } from '../../../../../db/knex-database-connection.js';
+import { _ } from '../../../../../lib/infrastructure/utils/lodash-utils.js';
 import { AttachableTargetProfile } from '../../domain/models/AttachableTargetProfile.js';
-import { _ } from '../utils/lodash-utils.js';
 
 const find = async function ({ searchTerm } = {}) {
   const targetProfiles = await knex('target-profiles')
@@ -10,7 +10,7 @@ const find = async function ({ searchTerm } = {}) {
     .leftJoin('complementary-certification-badges', 'badges.id', 'complementary-certification-badges.badgeId')
     .where('target-profiles.outdated', false)
     .where(_allowOnlyNeverAttachedTargetProfiles)
-    .where((builder) => _searchByCritieria({ builder, searchTerm }))
+    .where((builder) => _searchByCriteria({ builder, searchTerm }))
     .orderBy('target-profiles.name', 'ASC')
     .orderBy('target-profiles.id', 'DESC');
 
@@ -19,7 +19,7 @@ const find = async function ({ searchTerm } = {}) {
 
 export { find };
 
-function _searchByCritieria({ builder, searchTerm }) {
+function _searchByCriteria({ builder, searchTerm }) {
   if (!_.isBlank(searchTerm)) {
     const filteredBuilder = _searchByTargetProfileName({ builder, searchTerm });
     const isNumberOnly = /^\d+$/.test(searchTerm);
@@ -31,6 +31,7 @@ function _searchByCritieria({ builder, searchTerm }) {
   }
   return builder;
 }
+
 function _searchByTargetProfileName({ builder, searchTerm }) {
   return builder.whereILike('target-profiles.name', `%${searchTerm}%`);
 }
